Refactor exports, add tests for types
